cleanup routing handling
